Relocate shared utils and request options into next
Move the utils still needed by ./next (map classes, getMetricsForTelemetry,
getSignalForToken, isAuthenticationWwwAuthenticateRequest, Limiter, screen) and
the AzureSubscriptionProviderRequestOptions contract into the next folder,
re-exporting them from the legacy entrypoint. Rename tryGetTokenExpiration to
getTokenExpiry and mirror the moved util tests under test/next/utils. Have the
legacy configuredAzureEnv delegate to the next implementation, and turn the
pure dedupeSubscriptions wrapper into a direct re-export from next.
