+++ auth/src/next/contracts/AzureSubscriptionProviderRequestOptions.ts
