+++ auth/src/next/utils/getSignalForToken.ts
