+++ auth/src/next/utils/isAuthenticationWwwAuthenticateRequest.ts
