+++ auth/src/next/utils/screen.ts
@@ -3,8 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import type { AzureAccount } from '../contracts/AzureAccount';
-import type { AzureTenant } from '../contracts/AzureTenant';
+import type { AzureAccount } from '../../contracts/AzureAccount';
+import type { AzureTenant } from '../../contracts/AzureTenant';
 
 // These regexes are not perfect, but should work for common cases
 // If they don't work, we'll just return the account ID, which does not need to be screened
