+++ auth/src/providers/VSCodeAzureSubscriptionProvider.ts
@@ -9,7 +9,7 @@ import type { AzureAccount } from '../contracts/AzureAccount';
 import type { AzureAuthentication } from '../contracts/AzureAuthentication';
 import type { AzureSubscription } from '../contracts/AzureSubscription';
 import type { TenantIdAndAccount } from '../contracts/AzureSubscriptionProvider';
-import { type BaseOptions, DefaultOptions, type GetAccountsOptions, type GetAvailableSubscriptionsOptions, type GetSubscriptionsForTenantOptions, type GetTenantsForAccountOptions } from '../contracts/AzureSubscriptionProviderRequestOptions'; // eslint-disable-line @typescript-eslint/no-unused-vars -- It is used in the doc comments
+import { type BaseOptions, DefaultOptions, type GetAccountsOptions, type GetAvailableSubscriptionsOptions, type GetSubscriptionsForTenantOptions, type GetTenantsForAccountOptions } from '../next/contracts/AzureSubscriptionProviderRequestOptions'; // eslint-disable-line @typescript-eslint/no-unused-vars -- It is used in the doc comments
 import type { AzureTenant } from '../contracts/AzureTenant';
 import { createAzureLoggerForOutputChannel } from './azureLoggerForOutputChannel';
 import { buildLegacyAuthentication, projectAccount, projectSubscription, projectTenant } from './projectToLegacy';
