+++ auth/src/utils/configuredAzureEnv.ts
@@ -5,17 +5,8 @@
 
 import * as azureEnv from '@azure/ms-rest-azure-env'; // This package is so small that it's not worth lazy loading
 import * as vscode from 'vscode';
-
-// These strings come from https://github.com/microsoft/vscode/blob/eac16e9b63a11885b538db3e0b533a02a2fb8143/extensions/microsoft-authentication/package.json#L40-L99
-export const CustomCloudConfigurationSection = 'microsoft-sovereign-cloud';
-const CloudEnvironmentSettingName = 'environment';
-const CustomEnvironmentSettingName = 'customEnvironment';
-
-enum CloudEnvironmentSettingValue {
-    ChinaCloud = 'ChinaCloud',
-    USGovernment = 'USGovernment',
-    Custom = 'custom',
-}
+import { AzureChinaCloud, AzureUSGovernmentCloud } from '../next/contracts/EnvironmentLike';
+import { getConfiguredAuthProviderId as nextGetConfiguredAuthProviderId, getConfiguredAzureEnv as nextGetConfiguredAzureEnv, setConfiguredAzureEnv as nextSetConfiguredAzureEnv } from '../next/configuredEnvironment';
 
 export class ExtendedEnvironment extends azureEnv.Environment {
     public constructor(parameters: azureEnv.EnvironmentParameters, public readonly isCustomCloud: boolean) {
@@ -29,31 +20,34 @@ export class ExtendedEnvironment extends azureEnv.Environment {
 /**
  * Gets the configured Azure environment.
  *
+ * @remarks This delegates the configuration lookup to the dependency-injected `./next`
+ * {@link nextGetConfiguredAzureEnv}, then maps the structural {@link EnvironmentLike} back to an
+ * `@azure/ms-rest-azure-env`-based {@link ExtendedEnvironment} for backwards compatibility.
+ *
  * @returns The configured Azure environment from the settings in the built-in authentication provider extension
  */
 export function getConfiguredAzureEnv(): ExtendedEnvironment {
-    const authProviderConfig = vscode.workspace.getConfiguration(CustomCloudConfigurationSection);
-    const environmentSettingValue = authProviderConfig.get<string | undefined>(CloudEnvironmentSettingName);
+    const { environment, isCustomCloud } = nextGetConfiguredAzureEnv(vscode);
 
-    if (environmentSettingValue === CloudEnvironmentSettingValue.ChinaCloud) {
-        return new ExtendedEnvironment(azureEnv.Environment.ChinaCloud, false);
-    } else if (environmentSettingValue === CloudEnvironmentSettingValue.USGovernment) {
-        return new ExtendedEnvironment(azureEnv.Environment.USGovernment, false);
-    } else if (environmentSettingValue === CloudEnvironmentSettingValue.Custom) {
-        const customCloud = authProviderConfig.get<azureEnv.EnvironmentParameters | undefined>(CustomEnvironmentSettingName);
-
-        if (customCloud) {
-            return new ExtendedEnvironment(customCloud, true);
-        }
-
-        throw new Error(vscode.l10n.t('The custom cloud choice is not configured. Please configure the setting `{0}.{1}`.', CustomCloudConfigurationSection, CustomEnvironmentSettingName));
+    if (isCustomCloud) {
+        return new ExtendedEnvironment(environment, true);
     }
 
-    return new ExtendedEnvironment(azureEnv.Environment.AzureCloud, false);
+    switch (environment) {
+        case AzureChinaCloud:
+            return new ExtendedEnvironment(azureEnv.Environment.ChinaCloud, false);
+        case AzureUSGovernmentCloud:
+            return new ExtendedEnvironment(azureEnv.Environment.USGovernment, false);
+        default:
+            return new ExtendedEnvironment(azureEnv.Environment.AzureCloud, false);
+    }
 }
 
 /**
  * Sets the configured Azure cloud.
+ *
+ * @remarks This is a thin wrapper around the dependency-injected `./next` {@link nextSetConfiguredAzureEnv},
+ * binding it to the real `vscode` namespace.
  *
  * @param cloud Use `'AzureCloud'` or `undefined` for public Azure cloud, `'ChinaCloud'` for Azure China, or `'USGovernment'` for Azure US Government.
  * These are the same values as the cloud names in `@azure/ms-rest-azure-env`. For a custom cloud, use an instance of the `@azure/ms-rest-azure-env` {@link azureEnv.EnvironmentParameters}.
@@ -61,30 +55,17 @@ export function getConfiguredAzureEnv(): ExtendedEnvironment {
  * @param target (Optional) The configuration target to use, by default {@link vscode.ConfigurationTarget.Global}.
  */
 export async function setConfiguredAzureEnv(cloud: 'AzureCloud' | 'ChinaCloud' | 'USGovernment' | undefined | null | azureEnv.EnvironmentParameters, target: vscode.ConfigurationTarget = vscode.ConfigurationTarget.Global): Promise<void> {
-    const authProviderConfig = vscode.workspace.getConfiguration(CustomCloudConfigurationSection);
-
-    if (typeof cloud === 'undefined' || !cloud) {
-        // Use public cloud implicitly--set `environment` setting to `undefined`
-        await authProviderConfig.update(CloudEnvironmentSettingName, undefined, target);
-    } else if (typeof cloud === 'string' && cloud === 'AzureCloud') {
-        // Use public cloud explicitly--set `environment` setting to `undefined`
-        await authProviderConfig.update(CloudEnvironmentSettingName, undefined, target);
-    } else if (typeof cloud === 'string') {
-        // Use a sovereign cloud--set the `environment` setting to the specified value
-        await authProviderConfig.update(CloudEnvironmentSettingName, cloud, target);
-    } else if (typeof cloud === 'object') {
-        // use a custom cloud--set the `environment` setting to `custom` and the `customEnvironment` setting to the specified value
-        await authProviderConfig.update(CloudEnvironmentSettingName, CloudEnvironmentSettingValue.Custom, target);
-        await authProviderConfig.update(CustomEnvironmentSettingName, cloud, target);
-    } else {
-        throw new Error(`Invalid cloud value: ${JSON.stringify(cloud)}`);
-    }
+    await nextSetConfiguredAzureEnv(vscode, cloud, target);
 }
 
 /**
  * Gets the ID of the authentication provider configured to be used
+ *
+ * @remarks This is a thin wrapper around the dependency-injected `./next` {@link nextGetConfiguredAuthProviderId},
+ * binding it to the real `vscode` namespace.
+ *
  * @returns The provider ID to use, either `'microsoft'` or `'microsoft-sovereign-cloud'`
  */
 export function getConfiguredAuthProviderId(): string {
-    return getConfiguredAzureEnv().name === azureEnv.Environment.AzureCloud.name ? 'microsoft' : 'microsoft-sovereign-cloud';
+    return nextGetConfiguredAuthProviderId(vscode);
 }
